backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
python-decouple==3.8
celery==5.3.4
psycopg2-binary==2.9.9
Pillow==10.1.0
drf-spectacular==0.29.0
drf-spectacular==0.29.0